Rewrite key management (kid-aware, backup refresh, concurrent fetches)
Key management is surprisingly tricky. We want to cache the public keys
to not send an HTTP request to the JWKS endpoint on every incoming
request. But sometimes, when the JWT provider rotates keys, we want to
refresh the keys even though they are not technically stale yet, to
check if there is a new fitting one. This is "backup refresh" and it
is rate-limited.

This rewrite also makes everything `kid` aware, i.e. the key ID of JWTs
and JWKS is respected. It is still not required. Setting the kid can
make some lookup operations faster, but all in all I don't really see
its use yet. Especially since we want to keep JWTs small.

Finally, JWKS fetches can now happen in parallel for different sources.

The most tricky thing about this is to ensure no fetches for the same
URL happen at the same time, while still allowing all tasks to wait for
exactly what they need and not more. I am pretty sure I did it correctly
now and I think the code isn't too bad to understand.