Replace `on_forbidden = "proxy"` with `opencast.use_as_fallback`
The proxy approach had multiple problems, the main one being: the web
server (eg nginx) "including" octoka expects a certain response
(configured as `on_allow`). When using `proxy`, the same response is
still expected from octoka. But to achieve that, OC would need to be
configured in the same way as octoka to ensure that. And that's not even
fully possible because OC does not have an equivalent of
`on_success = "empty"`.

It makes much more sense for octoka to always craft its own responses
according to `on_allow`. OC is just asked as fallback, with octoka
treating it as a secondary auth check.

This commit reverts quite a few changes from the last one.