Reply 404 if asking Opencast returns 404
See comment in the diff for explanation. Another idea would be to treat
404 as authorized, as in OC, currently, the auth check is done before
checking whether the file exists. So if it would have been forbidden,
403 would have been returned. But OC could change this at any time, so
it's kind of icky?

Replying 404 is certainly the correct thing for "file" and "x-accel"
options of on_success. Only for the "empty" case, which is used for
nginx auth_request, it's questionable.