Change `use_as_fallback` to `fallback` and allow different modes
This is mainly to work around a bug in OC, where HEAD requests are
incorrectly answered. But it also opens up the option to add a dedicated
API to OC in the future and use that. HEAD requests still do a lot of
unnecessary work and have a problem with 404 responses.