Use jemalloc to improve performance a bit
The loadtest of the last commit showed these numbers for musl:

|         | default alloc | jemalloc | linux-gnu |
| ------- | ------------- | -------- | --------- |
| empty   | 111k          | 122k     | 121k      |
| file    |  32k          |  44k     |  45k      |

So the benefits are not as large as they were for Tobira, for example.
But especially in the "file" case, it's still quite a difference, and
getting this performance boost is basically free (just add the few
lines in this commit).