Fix behavior of `prefix` header JWT source (make prefix required)
Before, if the header did not start with the prefix, that was ignored
and the full header value was considered. This was a bug, as it's not
intuitive nor helpful. With this lenient behavior, HTTP digest auth
for example always triggers parsing `Digest realm=...` as a JWT, which
obviously fails. So this was not a security relevant bug, but just
resulted in needlessly parsing JWTs and confusing log outputs.